Add database
Added mongodb database. Added models, controllers, and routes folders. Added blogController.js, blogRouter.js, and Blog.js.

Blog.js creates the schema for blog posts. The Blog schema is imported into the blogController. The blogController defines one method, createBlogPost that takes the request body and creates a new blog post in the database. Once the blog post is created, it redirects the user to "/".

The createBlogPost method is imported into blogRoutes.js where it is used as middleware on the /new-blog-post route. The blogRoutes.js is imported to app.js as blogRouter and is used on the "api/v1/blogs" route.

Updated index.ejs so when the user submits a new blog post it goes to "api/v1/blogs/new-blog-post".

We can now sucessfully add posts to the MongoDB database!