Refactor to use db instead of blogPost array
Refactored app so we can:
i) get the blog posts from the database and display them
ii) add a blog post to the database and rerender the view for the user.

Pagination now works with the database.

Moved variables needed globally into a blogConfig.js file in the config folder.

Added helpers folder and a blogHelpers.js file. Added getBlogCount method that retrieves the total blog post count from the database.

Added route "/" to blogRoutes that uses a GET request and utilizes getBlogPostsForCurrentPage controller.

Decided not to use "api/v1/blogs/ as a route as our server will only be used by our index.ejs frontend. Blog routes are now on "/" to get, and we can add other routes for other crud operations as needed - e.g. /new-blog-post to create a blog post.

Added getBlogPostsForCurrentPage to blogController. We use the query.pagenum to determine what posts to get. We then retrieve the posts we need from the database using skip and limit, and get the total posts and then render the index.ejs for the user.